Break out utility function to consolidate duplicated logic
