+++ codemods/src/6.0.0-7.0.0/replace-classic-theme-imports.ts
@@ -1,6 +1,5 @@
 import { Transform } from 'jscodeshift'
-import { getLogger } from '../utils/get-logger'
-import { registerExtensions } from '../utils/register-extensions'
+import { replaceEvergreenImportWithLocalPath } from '../utils/replace-evergreen-import-with-local-path'
 
 /**
  * Codemod to replace imports referencing the `classicTheme` object from evergreen-ui to a provided path
@@ -9,52 +8,7 @@ import { registerExtensions } from '../utils/register-extensions'
  * ```
  */
 
-const CLASSIC_THEME = 'classicTheme'
-
-const transformer: Transform = (file, api, options) => {
-  const log = getLogger(file)
-
-  const { localPath } = options
-  if (localPath == null || localPath.length < 1) {
-    log(
-      `Local path is required for replacement. Specify where the ${CLASSIC_THEME} object should be imported from, i.e. --localPath=themes/classic-theme`
-    )
-    return file.source
-  }
-
-  const j = registerExtensions(api.jscodeshift)
-
-  const root = j(file.source)
-
-  const importDeclarations = root.findEvergreenImportDeclaration()
-
-  if (importDeclarations.isEmpty()) {
-    return file.source
-  }
-
-  const importSpecifiers = importDeclarations.findImportSpecifiersByName(CLASSIC_THEME)
-
-  if (importSpecifiers.isEmpty()) {
-    return file.source
-  }
-
-  // Save reference to original importSpecifier to swap it with the local path (i.e. any alias that might be used)
-  const importSpecifier = importSpecifiers.firstNode()!
-
-  // Remove the original evergreen import
-  importSpecifiers.remove()
-
-  const hasLocalThemeImport = root
-    .findImportDeclarationByModuleName(localPath)
-    .findImportSpecifiersByName(CLASSIC_THEME)
-    .hasValues()
-
-  if (!hasLocalThemeImport) {
-    // Add an import to the local path using the saved import specifier
-    importDeclarations.insertAfter(j.importDeclaration([j.importSpecifier.from(importSpecifier)], j.literal(localPath)))
-  }
-
-  return root.toSource()
-}
+const transformer: Transform = (file, api, options) =>
+  replaceEvergreenImportWithLocalPath({ file, api, localPath: options.localPath, importName: 'classicTheme' })
 
 export default transformer
