+++ codemods/src/utils/register-import-declaration-collection-extensions.ts
@@ -33,7 +33,7 @@ const _registerImportDeclarationCollectionExtensions = (
         const importDeclarations = (this as any) as Collection<ImportDeclaration>
         return (importDeclarations.find(
           j.ImportSpecifier,
-          importSpecifier => importSpecifier.local?.name === name || importSpecifier.imported.name === name
+          importSpecifier => importSpecifier.imported.name === name
         ) as any) as ExtendedImportSpecifierCollection
       }
     },
